requirements.txt
fastapi==0.115.13
uvicorn[standard]==0.32.1
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.66
langchain-text-splitters==0.3.8
openai==1.91.0
chromadb==1.0.13
python-dotenv==1.0.1
pydantic==2.10.4